requirements.txt
aio-pika==9.4.2
aiormq==6.8.0
annotated-types==0.7.0
async-timeout==4.0.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
iniconfig==2.0.0
multidict==6.0.5
packaging==24.1
pamqp==3.3.0
pika==1.3.2
playwright==1.45.1
pluggy==1.5.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
pyparsing==3.1.2
pytest==8.3.2
pytest-base-url==2.1.0
pytest-playwright==0.5.1
python-dotenv==1.0.1
python-slugify==8.0.4
ratelimit==2.2.1
redis==5.0.8
requests==2.32.3
rsa==4.9
tenacity==9.0.0
text-unidecode==1.3
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
